+++ funkgateway/__init__.py
@@ -3,4 +3,4 @@
 Open-source radio gateway software for Linux.
 SPDX-License-Identifier: GPL-3.0-or-later
 """
-__version__ = "0.5.9.44"
+__version__ = "0.5.9.45"